requirements.txt
PyYAML==6.0.1
requests==2.32.3
cmake==4.0.3
lcov
genhtml